refactored the navbar
