+++ src/containers/App.jsx
@@ -1,17 +1,7 @@
 import React, { Component, PropTypes } from 'react'
+import DictyNavbar from 'components/DictyNavbar'
 import { connect } from 'react-redux'
 import { StyleRoot } from 'radium'
-import Navbar from 'dicty-react-components/src/Navbar'
-import NavItem from 'dicty-react-components/src/NavItem'
-import NavbarHeader from 'dicty-react-components/src/NavbarHeader'
-import NavbarItems from 'dicty-react-components/src/NavbarItems'
-import NavbarDropdown from 'dicty-react-components/src/NavbarDropdown'
-import DropdownMenu from 'dicty-react-components/src/DropdownMenu'
-import Header from 'dicty-react-components/src/Header'
-import HeaderLogo from 'dicty-react-components/src/HeaderLogo'
-import HeaderItems from 'dicty-react-components/src/HeaderItems'
-import HeaderLinks from 'dicty-react-components/src/HeaderLinks'
-import SearchBox from 'dicty-react-components/src/SearchBox'
 import { bindActionCreators } from 'redux'
 import * as authActionCreators from 'actions/auth'
 import * as shippingActionCreators from 'actions/order/shipping'
@@ -32,40 +22,6 @@ class App extends Component {
         })
     };
     render() {
-        const navStyle = {
-            backgroundColor: '#1F4484',
-            border: '1px solid #1F4484',
-            marginTop: '10px'
-        }
-        const itemStyle = {
-            color: '#ffffff',
-            fontSize: '18px',
-            ':hover': {
-                color: '#ffffff',
-                backgroundColor: '#1b3c74'
-            },
-            ':focus': {
-                color: '#ffffff',
-                backgroundColor: '#1b3c74'
-            }
-        }
-        const menuItemStyle = {
-            ':hover': {
-                color: '#ffffff',
-                backgroundColor: '#1F4484'
-            },
-            ':focus': {
-                color: '#ffffff',
-                backgroundColor: '#1F4484'
-            },
-            '@media (max-width: 767px)': {
-                color: '#ffffff',
-
-                ':hover': {
-                    color: '#2574A9'
-                }
-            }
-        }
         const genomes = [
             {href: '#', name: 'Home'}
         ]
@@ -84,54 +40,16 @@ class App extends Component {
             {href: '#', name: 'Cite Us'},
             {href: '#', name: 'Dicty Annual Conferences'}
         ]
-        const headerLinks = [
-            {href: '#', name: 'Cite Us', iconClass: 'fa fa-book'},
-            {href: '#', name: 'Downloads', iconClass: 'fa fa-download'},
-            {href: '#', name: 'About Us', iconClass: 'fa fa-info-circle'}
-        ]
         return (
             <div>
                 <StyleRoot>
-                    <Header>
-                        <HeaderLogo path= "dicty-logo.png"
-                            href= "http://dictybase.org/"
-                        />
-                        <HeaderItems>
-                            <HeaderLinks links= { headerLinks } />
-                            <SearchBox />
-                        </HeaderItems>
-                    </Header>
-                    <Navbar navStyle={ navStyle }>
-                        <NavbarHeader href="/" name="dictyBase"/>
-                        <NavbarItems>
-                            <NavbarDropdown name="Genomes" itemStyle={ itemStyle }>
-                                <DropdownMenu menuItems={ genomes }
-                                    menuItemStyle={ menuItemStyle }
-                                />
-                            </NavbarDropdown>
-                            <NavbarDropdown name="Tools" itemStyle={ itemStyle }>
-                                <DropdownMenu menuItems={ tools }
-                                    menuItemStyle={ menuItemStyle }
-                                />
-                            </NavbarDropdown>
-                            <NavbarDropdown name="Explore" itemStyle={ itemStyle }>
-                                <DropdownMenu menuItems={ explore }
-                                    menuItemStyle={ menuItemStyle }
-                                />
-                            </NavbarDropdown>
-                            <NavbarDropdown name="Research" itemStyle={ itemStyle }>
-                                <DropdownMenu menuItems={ research }
-                                    menuItemStyle={ menuItemStyle }
-                                />
-                            </NavbarDropdown>
-                            <NavItem link="/" title="Dicty Stock Center" itemStyle={ itemStyle }/>
-                            <NavbarDropdown name="Community" itemStyle={ itemStyle }>
-                                <DropdownMenu menuItems={ community }
-                                    menuItemStyle={ menuItemStyle }
-                                />
-                            </NavbarDropdown>
-                        </NavbarItems>
-                    </Navbar>
+                    <DictyNavbar
+                        genomes={ genomes }
+                        tools={ tools }
+                        explore={ explore }
+                        research={ research }
+                        community={ community }
+                    />
                 </StyleRoot>
             { this.renderChildren() }
             </div>
